tigress-tool/requirements.txt
pyyaml
types-pyyaml
